Fix raising cache update events when a view starts up
