+++ src/GitHub.Api/Cache/CacheContainer.cs
@@ -1,5 +1,6 @@
 ﻿using System;
 using System.Collections.Generic;
+using System.Globalization;
 using GitHub.Logging;
 
 namespace GitHub.Unity
@@ -47,11 +48,11 @@ namespace GitHub.Unity
             return caches[cacheType].Value;
         }
 
-        public void CheckAndRaiseEventsIfCacheNewer(CacheUpdateEvent cacheUpdateEvent)
+        public void CheckAndRaiseEventsIfCacheNewer(CacheType cacheType, CacheUpdateEvent cacheUpdateEvent)
         {
-            var cache = GetCache(cacheUpdateEvent.cacheType);
+            var cache = GetCache(cacheType);
             var needsInvalidation = cache.ValidateData();
-            if (!needsInvalidation || cache.LastUpdatedAt != cacheUpdateEvent.UpdatedTime)
+            if (!cacheUpdateEvent.IsInitialized || !needsInvalidation || cache.LastUpdatedAt != cacheUpdateEvent.UpdatedTime)
             {
                 OnCacheUpdated(cache.CacheType, cache.LastUpdatedAt);
             }
@@ -98,5 +99,90 @@ namespace GitHub.Unity
         public IGitLocksCache GitLocksCache { get { return (IGitLocksCache)caches[CacheType.GitLocks].Value; } }
         public IGitUserCache GitUserCache { get { return (IGitUserCache)caches[CacheType.GitUser].Value; } }
         public IRepositoryInfoCache RepositoryInfoCache { get { return (IRepositoryInfoCache)caches[CacheType.RepositoryInfo].Value; } }
+    }
+
+    [Serializable]
+    public struct CacheUpdateEvent
+    {
+        [NonSerialized] private DateTimeOffset? updatedTimeValue;
+        public string updatedTimeString;
+        public CacheType cacheType;
+
+        public CacheUpdateEvent(CacheType type, DateTimeOffset when)
+        {
+            if (type == CacheType.None) throw new ArgumentOutOfRangeException(nameof(type));
+
+            cacheType = type;
+            updatedTimeValue = when;
+            updatedTimeString = when.ToString(Constants.Iso8601Format);
+        }
+
+        public override int GetHashCode()
+        {
+            int hash = 17;
+            hash = hash * 23 + cacheType.GetHashCode();
+            hash = hash * 23 + (updatedTimeString?.GetHashCode() ?? 0);
+            return hash;
+        }
+
+        public override bool Equals(object other)
+        {
+            if (other is CacheUpdateEvent)
+                return Equals((CacheUpdateEvent)other);
+            return false;
+        }
+
+        public bool Equals(CacheUpdateEvent other)
+        {
+            return
+                cacheType == other.cacheType &&
+                String.Equals(updatedTimeString, other.updatedTimeString)
+                ;
+        }
+
+        public static bool operator ==(CacheUpdateEvent lhs, CacheUpdateEvent rhs)
+        {
+            // If both are null, or both are same instance, return true.
+            if (ReferenceEquals(lhs, rhs))
+                return true;
+
+            // If one is null, but not both, return false.
+            if (((object)lhs == null) || ((object)rhs == null))
+                return false;
+
+            // Return true if the fields match:
+            return lhs.Equals(rhs);
+        }
+
+        public static bool operator !=(CacheUpdateEvent lhs, CacheUpdateEvent rhs)
+        {
+            return !(lhs == rhs);
+        }
+
+        public DateTimeOffset UpdatedTime
+        {
+            get
+            {
+                if (!updatedTimeValue.HasValue)
+                {
+                    DateTimeOffset result;
+                    if (DateTimeOffset.TryParseExact(updatedTimeString, Constants.Iso8601Format, CultureInfo.InvariantCulture, DateTimeStyles.None, out result))
+                    {
+                        updatedTimeValue = result;
+                    }
+                    else
+                    {
+                        updatedTimeValue = DateTimeOffset.MinValue;
+                        updatedTimeString = updatedTimeValue.Value.ToString(Constants.Iso8601Format);
+                    }
+                }
+
+                return updatedTimeValue.Value;
+            }
+        }
+
+        public bool IsInitialized => cacheType != CacheType.None;
+
+        public string UpdatedTimeString => updatedTimeString;
     }
 }
