+++ src/GitHub.Api/Git/Repository.cs
@@ -120,7 +120,7 @@ namespace GitHub.Unity
         public ITask ReleaseLock(string file, bool force) => repositoryManager.UnlockFile(file, force);
         public ITask DiscardChanges(GitStatusEntry[] gitStatusEntry) => repositoryManager.DiscardChanges(gitStatusEntry);
 
-        public void CheckAndRaiseEventsIfCacheNewer(CacheUpdateEvent cacheUpdateEvent) => cacheContainer.CheckAndRaiseEventsIfCacheNewer(cacheUpdateEvent);
+        public void CheckAndRaiseEventsIfCacheNewer(CacheType cacheType, CacheUpdateEvent cacheUpdateEvent) => cacheContainer.CheckAndRaiseEventsIfCacheNewer(cacheType, cacheUpdateEvent);
 
 
         /// <summary>
@@ -405,17 +405,7 @@ namespace GitHub.Unity
             cacheContainer.CacheUpdated += (type, dt) => { if (type == CacheType.GitUser) CacheHasBeenUpdated(dt); };
         }
 
-        public void CheckUserChangedEvent(CacheUpdateEvent cacheUpdateEvent)
-        {
-            var managedCache = cacheContainer.GitUserCache;
-            var raiseEvent = !cacheUpdateEvent.IsInitialized || managedCache.LastUpdatedAt != cacheUpdateEvent.UpdatedTime;
-
-            Logger.Trace("Check GitUserCache CacheUpdateEvent Current:{0} Check:{1} Result:{2}", managedCache.LastUpdatedAt,
-                cacheUpdateEvent.UpdatedTime, raiseEvent);
-
-            if (raiseEvent)
-                CacheHasBeenUpdated(managedCache.LastUpdatedAt);
-        }
+        public void CheckUserChangedEvent(CacheUpdateEvent cacheUpdateEvent) => cacheContainer.CheckAndRaiseEventsIfCacheNewer(CacheType.GitUser, cacheUpdateEvent);
 
         public void Initialize(IGitClient client)
         {
@@ -492,90 +482,5 @@ namespace GitHub.Unity
         }
 
         protected static ILogging Logger { get; } = LogHelper.GetLogger<User>();
-    }
-
-    [Serializable]
-    public struct CacheUpdateEvent
-    {
-        [NonSerialized] private DateTimeOffset? updatedTimeValue;
-        public string updatedTimeString;
-        public CacheType cacheType;
-
-        public CacheUpdateEvent(CacheType type, DateTimeOffset when)
-        {
-            if (type == CacheType.None) throw new ArgumentOutOfRangeException(nameof(type));
-
-            cacheType = type;
-            updatedTimeValue = when;
-            updatedTimeString = when.ToString(Constants.Iso8601Format);
-        }
-
-        public override int GetHashCode()
-        {
-            int hash = 17;
-            hash = hash * 23 + cacheType.GetHashCode();
-            hash = hash * 23 + (updatedTimeString?.GetHashCode() ?? 0);
-            return hash;
-        }
-
-        public override bool Equals(object other)
-        {
-            if (other is CacheUpdateEvent)
-                return Equals((CacheUpdateEvent)other);
-            return false;
-        }
-
-        public bool Equals(CacheUpdateEvent other)
-        {
-            return
-                cacheType == other.cacheType &&
-                String.Equals(updatedTimeString, other.updatedTimeString)
-                ;
-        }
-
-        public static bool operator ==(CacheUpdateEvent lhs, CacheUpdateEvent rhs)
-        {
-            // If both are null, or both are same instance, return true.
-            if (ReferenceEquals(lhs, rhs))
-                return true;
-
-            // If one is null, but not both, return false.
-            if (((object)lhs == null) || ((object)rhs == null))
-                return false;
-
-            // Return true if the fields match:
-            return lhs.Equals(rhs);
-        }
-
-        public static bool operator !=(CacheUpdateEvent lhs, CacheUpdateEvent rhs)
-        {
-            return !(lhs == rhs);
-        }
-
-        public DateTimeOffset UpdatedTime
-        {
-            get
-            {
-                if (!updatedTimeValue.HasValue)
-                {
-                    DateTimeOffset result;
-                    if (DateTimeOffset.TryParseExact(updatedTimeString, Constants.Iso8601Format, CultureInfo.InvariantCulture, DateTimeStyles.None, out result))
-                    {
-                        updatedTimeValue = result;
-                    }
-                    else
-                    {
-                        updatedTimeValue = DateTimeOffset.MinValue;
-                        updatedTimeString = updatedTimeValue.Value.ToString(Constants.Iso8601Format);
-                    }
-                }
-
-                return updatedTimeValue.Value;
-            }
-        }
-
-        public bool IsInitialized => cacheType != CacheType.None;
-
-        public string UpdatedTimeString => updatedTimeString;
     }
 }
