+++ src/UnityExtension/Assets/Editor/GitHub.Unity/UI/SettingsView.cs
@@ -158,13 +158,8 @@ namespace GitHub.Unity
 
         private void ValidateCachedData(IRepository repository)
         {
-            if (lastCurrentRemoteChangedEvent.cacheType == CacheType.None)
-                lastCurrentRemoteChangedEvent = new CacheUpdateEvent(CacheType.RepositoryInfo, DateTimeOffset.MinValue);
-            if (lastLocksChangedEvent.cacheType == CacheType.None)
-                lastLocksChangedEvent = new CacheUpdateEvent(CacheType.GitLocks, DateTimeOffset.MinValue);
-
-            repository.CheckAndRaiseEventsIfCacheNewer(lastCurrentRemoteChangedEvent);
-            repository.CheckAndRaiseEventsIfCacheNewer(lastLocksChangedEvent);
+            repository.CheckAndRaiseEventsIfCacheNewer(CacheType.RepositoryInfo, lastCurrentRemoteChangedEvent);
+            repository.CheckAndRaiseEventsIfCacheNewer(CacheType.GitLocks, lastLocksChangedEvent);
         }
 
         private void MaybeUpdateData()
